Change the definition of TargetRegisterInfo::getCrossCopyRegClass to be more flexible.
If it returns a register class that's different from the input, then that's the
register class used for cross-register class copies.
If it returns a register class that's the same as the input, then no cross-
register class copies are needed (normal copies would do).
If it returns null, then it's not at all possible to copy registers of the
specified register class.


git-svn-id: https://llvm.org/svn/llvm-project@127368 91177308-0d34-0410-b5e6-96231b3b80d8