Free isl_ctx and fix several memory leaks
Because of me not understanding the LLVM pass structure well, I did not find a
good way to allocate isl_ctx and to free it later without getting issues with
reference counting. I now found this place, such that we can free isl_ctx. This
patch also fixes the memory leaks that were ignored beforehand.

git-svn-id: https://llvm.org/svn/llvm-project@138204 91177308-0d34-0410-b5e6-96231b3b80d8