Reapply r138984 with changes suggested by Duncan:
Update to use the new EH scheme.

This uses the landingpad instruction in place of the llvm.eh.exception and
llvm.eh.selector calls. The values returned by the instruction are always
stored. The 'resume' instruction replaces the direct call to _Unwind_Resume.

This is based partially off of dragonegg's EH implementation.


git-svn-id: https://llvm.org/svn/llvm-project@139212 91177308-0d34-0410-b5e6-96231b3b80d8